LayoutTests/platform/ios-simulator/ios/fast/forms/autocapitalize-attribute-expected.txt

Test the "autocapitalize" attribute values and fallback values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing inputs outside of a form.
PASS i1.autocapitalize is ""
PASS i1.getAttribute('autocapitalize') is null
PASS i2.autocapitalize is ""
PASS i2.getAttribute('autocapitalize') is ""
PASS i3.autocapitalize is "sentences"
PASS i3.getAttribute('autocapitalize') is "on"
PASS i4.autocapitalize is "none"
PASS i4.getAttribute('autocapitalize') is "off"
PASS i5.autocapitalize is "none"
PASS i5.getAttribute('autocapitalize') is "none"
PASS i6.autocapitalize is "words"
PASS i6.getAttribute('autocapitalize') is "words"
PASS i7.autocapitalize is "sentences"
PASS i7.getAttribute('autocapitalize') is "sentences"
PASS i8.autocapitalize is "characters"
PASS i8.getAttribute('autocapitalize') is "characters"
PASS i9.autocapitalize is "sentences"
PASS i9.getAttribute('autocapitalize') is "unknown"
PASS i10.autocapitalize is "words"
PASS i10.getAttribute('autocapitalize') is "wOrDs"

Testing forms.
PASS form1.autocapitalize is ""
PASS form1.getAttribute('autocapitalize') is null
PASS form2.autocapitalize is ""
PASS form2.getAttribute('autocapitalize') is ""
PASS form3.autocapitalize is "sentences"
PASS form3.getAttribute('autocapitalize') is "on"
PASS form4.autocapitalize is "words"
PASS form4.getAttribute('autocapitalize') is "wOrDs"
PASS form5.autocapitalize is "characters"
PASS form5.getAttribute('autocapitalize') is "characters"

Testing inputs inside of a form with a default of "characters".
PASS j1.autocapitalize is "characters"
PASS j1.getAttribute('autocapitalize') is null
PASS j2.autocapitalize is "characters"
PASS j2.getAttribute('autocapitalize') is ""
PASS j3.autocapitalize is "sentences"
PASS j3.getAttribute('autocapitalize') is "on"
PASS j4.autocapitalize is "none"
PASS j4.getAttribute('autocapitalize') is "off"
PASS j5.autocapitalize is "none"
PASS j5.getAttribute('autocapitalize') is "none"
PASS j6.autocapitalize is "words"
PASS j6.getAttribute('autocapitalize') is "words"
PASS j7.autocapitalize is "sentences"
PASS j7.getAttribute('autocapitalize') is "sentences"
PASS j8.autocapitalize is "characters"
PASS j8.getAttribute('autocapitalize') is "characters"
PASS j9.autocapitalize is "sentences"
PASS j9.getAttribute('autocapitalize') is "unknown"
PASS j10.autocapitalize is "words"
PASS j10.getAttribute('autocapitalize') is "wOrDs"

Testing inputs inside of a form with no default.
PASS x1.autocapitalize is ""
PASS x1.getAttribute('autocapitalize') is null
PASS x2.autocapitalize is ""
PASS x2.getAttribute('autocapitalize') is ""

Testing textareas outside of a form.
PASS text_i1.autocapitalize is ""
PASS text_i1.getAttribute('autocapitalize') is null
PASS text_i2.autocapitalize is ""
PASS text_i2.getAttribute('autocapitalize') is ""
PASS text_i3.autocapitalize is "sentences"
PASS text_i3.getAttribute('autocapitalize') is "on"
PASS text_i4.autocapitalize is "none"
PASS text_i4.getAttribute('autocapitalize') is "off"
PASS text_i5.autocapitalize is "none"
PASS text_i5.getAttribute('autocapitalize') is "none"
PASS text_i6.autocapitalize is "words"
PASS text_i6.getAttribute('autocapitalize') is "words"
PASS text_i7.autocapitalize is "sentences"
PASS text_i7.getAttribute('autocapitalize') is "sentences"
PASS text_i8.autocapitalize is "characters"
PASS text_i8.getAttribute('autocapitalize') is "characters"
PASS text_i9.autocapitalize is "sentences"
PASS text_i9.getAttribute('autocapitalize') is "unknown"
PASS text_i10.autocapitalize is "words"
PASS text_i10.getAttribute('autocapitalize') is "wOrDs"

Testing textareas inside of a form with a default of "characters".
PASS text_j1.autocapitalize is "characters"
PASS text_j1.getAttribute('autocapitalize') is null
PASS text_j2.autocapitalize is "characters"
PASS text_j2.getAttribute('autocapitalize') is ""
PASS text_j3.autocapitalize is "sentences"
PASS text_j3.getAttribute('autocapitalize') is "on"
PASS text_j4.autocapitalize is "none"
PASS text_j4.getAttribute('autocapitalize') is "off"
PASS text_j5.autocapitalize is "none"
PASS text_j5.getAttribute('autocapitalize') is "none"
PASS text_j6.autocapitalize is "words"
PASS text_j6.getAttribute('autocapitalize') is "words"
PASS text_j7.autocapitalize is "sentences"
PASS text_j7.getAttribute('autocapitalize') is "sentences"
PASS text_j8.autocapitalize is "characters"
PASS text_j8.getAttribute('autocapitalize') is "characters"
PASS text_j9.autocapitalize is "sentences"
PASS text_j9.getAttribute('autocapitalize') is "unknown"
PASS text_j10.autocapitalize is "words"
PASS text_j10.getAttribute('autocapitalize') is "wOrDs"

Testing textareas inside of a form with no default.
PASS text_x1.autocapitalize is ""
PASS text_x1.getAttribute('autocapitalize') is null
PASS text_x2.autocapitalize is ""
PASS text_x2.getAttribute('autocapitalize') is ""
PASS successfullyParsed is true

TEST COMPLETE
